big safty refactor, includes bottle of perfume
