+++ src/a06_bud_logic/_utils/str_helpers.py
@@ -1,0 +1,43 @@
+def budunit_str() -> str:
+    return "budunit"
+
+
+def bud_acctunit_str() -> str:
+    return "bud_acctunit"
+
+
+def bud_acct_membership_str() -> str:
+    return "bud_acct_membership"
+
+
+def bud_itemunit_str() -> str:
+    return "bud_itemunit"
+
+
+def bud_item_awardlink_str() -> str:
+    return "bud_item_awardlink"
+
+
+def bud_item_reasonunit_str() -> str:
+    return "bud_item_reasonunit"
+
+
+def bud_item_reason_premiseunit_str() -> str:
+    return "bud_item_reason_premiseunit"
+
+
+def bud_item_teamlink_str() -> str:
+    return "bud_item_teamlink"
+
+
+def bud_item_healerlink_str() -> str:
+    return "bud_item_healerlink"
+
+
+def bud_item_factunit_str() -> str:
+    return "bud_item_factunit"
+
+
+def bud_groupunit_str() -> str:
+    """Is not saved to raw data budunit json"""
+    return "bud_groupunit"
