+++ src/a06_bud_logic/test_bud_tool/test_bud_get_obj.py
@@ -10,6 +10,8 @@ from src.a06_bud_logic.bud_tool import (
     bud_item_reason_premiseunit_get_obj as premiseunit_get_obj,
     bud_item_factunit_get_obj,
     bud_get_obj,
+)
+from src.a06_bud_logic._utils.str_helpers import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
