+++ src/a09_pack_logic/_utils/example_deltas.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic.bud_tool import budunit_str, bud_acctunit_str
+from src.a06_bud_logic._utils.str_helpers import budunit_str, bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import atom_delete, atom_update, acct_name_str
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a09_pack_logic.delta import BudDelta, buddelta_shop
