+++ src/a09_pack_logic/legible.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.dict_toolbox import get_from_nested_dict
 from src.a06_bud_logic.bud import BudUnit
-from src.a06_bud_logic.bud_tool import (
+from src.a06_bud_logic._utils.str_helpers import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
