+++ src/a09_pack_logic/test_delta/test_delta_legible_acct_membership.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic.bud_tool import bud_acct_membership_str
+from src.a06_bud_logic._utils.str_helpers import bud_acct_membership_str
 from src.a08_bud_atom_logic.atom import (
     budatom_shop,
     atom_update,
