+++ src/a09_pack_logic/test_delta/test_delta_legible_item_teamlink.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic.bud_tool import bud_item_teamlink_str
+from src.a06_bud_logic._utils.str_helpers import bud_item_teamlink_str
 from src.a08_bud_atom_logic.atom_config import (
     atom_insert,
     atom_delete,
