+++ src/a10_bud_calc/test_/test_bud_calc_config.py
@@ -5,7 +5,7 @@ from src.a02_finance_logic._utils.str_helpers import (
     owner_name_str,
     world_id_str,
 )
-from src.a06_bud_logic.bud_tool import (
+from src.a06_bud_logic._utils.str_helpers import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
@@ -18,7 +18,6 @@ from src.a06_bud_logic.bud_tool import (
     bud_item_factunit_str,
     bud_groupunit_str,
 )
-
 from src.a08_bud_atom_logic.atom_config import (
     get_atom_config_dict,
     acct_name_str,
