+++ src/a17_idea_logic/test_idea/test_idea_column.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic.bud_tool import bud_acctunit_str
+from src.a06_bud_logic._utils.str_helpers import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     acct_name_str,
     group_label_str,
