+++ src/a17_idea_logic/test_idea/test_idea_format_builder.py
@@ -1,6 +1,8 @@
 from src.a00_data_toolbox.file_toolbox import save_json
 from src.a02_finance_logic._utils.str_helpers import owner_name_str, fisc_tag_str
-from src.a06_bud_logic.bud_tool import bud_itemunit_str
+from src.a06_bud_logic._utils.str_helpers import bud_itemunit_str
+
+# from src.a06_bud_logic._utils.str_helpers import
 from src.a08_bud_atom_logic.atom_config import (
     parent_road_str,
     item_tag_str,
@@ -13,6 +15,7 @@ from src.a17_idea_logic.idea_config import (
     dimens_str,
     attributes_str,
 )
+from src.a06_bud_logic._utils.str_helpers import bud_itemunit_str
 
 
 def create_dimens_idea_format_dict() -> dict:
