+++ src/a17_idea_logic/test_idea/test_idea_format_config.py
@@ -1,7 +1,6 @@
-from src.a00_data_toolbox.dict_toolbox import create_sorted_concatenated_str
 from src.a00_data_toolbox.file_toolbox import get_dir_file_strs, create_path
 from src.a02_finance_logic._utils.str_helpers import owner_name_str, fisc_tag_str
-from src.a06_bud_logic.bud_tool import bud_acctunit_str
+from src.a06_bud_logic._utils.str_helpers import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     face_name_str,
     event_int_str,
