+++ src/a19_world_logic/test_06_create_guts/test_world_etl06_03_bud_agg_tables_to_fisc_events_csvs.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.file_toolbox import create_path, open_file
 from src.a02_finance_logic._utils.str_helpers import owner_name_str, fisc_tag_str
-from src.a06_bud_logic.bud_tool import bud_acctunit_str
+from src.a06_bud_logic._utils.str_helpers import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     face_name_str,
     acct_name_str,
